# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alexander Evan API",
        "description": "Dokumentasi API Web Untuk Belajar",
        "contact": {
            "name": "Energeek",
            "email": "energeekmail@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base URL"
        }
    ],
    "paths": {
        "/api/candidates": {
            "get": {
                "tags": [
                    "API|MASTER|INDEX CANDIDATES"
                ],
                "summary": "Candidates index",
                "operationId": "67b4b336050bec33b41c72f9d4c0c97b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/candidates/store/": {
            "post": {
                "tags": [
                    "Api|Master|STORE CANDIDATES"
                ],
                "summary": "Jobs store",
                "operationId": "3e90293e7f012eca8b8dee29bbeb9d5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "telepon",
                                    "year",
                                    "skill_id",
                                    "job_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telepon": {
                                        "type": "number"
                                    },
                                    "year": {
                                        "type": "number"
                                    },
                                    "skill_id": {
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/candidates/show/{id}": {
            "get": {
                "tags": [
                    "API|MASTER|SHOW CANDIDATES"
                ],
                "summary": "Candidates show",
                "operationId": "47729c36256d967c2de97b2f3793bf43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/candidates/update/{id}": {
            "put": {
                "tags": [
                    "Api|MASTER|UPDATE CANDIDATES"
                ],
                "summary": "Jobs update",
                "operationId": "61aac852036c5a2e9374b34d6aa4821a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "telepon",
                                    "year",
                                    "skill_id",
                                    "job_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "telepon": {
                                        "type": "number"
                                    },
                                    "year": {
                                        "type": "number"
                                    },
                                    "skill_id": {
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/candidates/delete/{id}": {
            "delete": {
                "tags": [
                    "Api|Master|DELETE CANDIDATES"
                ],
                "summary": "Candidates Delete",
                "operationId": "2b9f433a243b6121d43e93cdcd4007db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "responseFormatter": {
                "properties": {
                    "status": {
                        "properties": {
                            "code": {
                                "type": "integer"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "object",
                        "default": null
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "globalResponse": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseFormatter"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "authBearerToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "authBearerToken": []
        }
    ]
}